Skeleton for repo selector
