+++ ui/src/repo/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html>
+<head>
+	<meta charset="utf8">
+	<meta name="viewport" content="width=device-width">
+
+	<title>prvt</title>
+</head>
+
+<body class="bg-shade-base text-text-base tracking-wider leading-normal">
+</body>
+
+</html>
