+++ ui/src/shared/components/ThemePicker.svelte
